src/requirements.txt
numpy==1.22.3
pillow==9.2.0
pycocotools==2.0.4
pydevd_pycharm==222.4167.4
pyyaml==6.0
torch==1.11.0+cu113
torchvision==0.12.0+cu113
wandb==0.12.18
